Latest changes in rotational motion state derivatives: Added new rotational propagators to rotational settings Fixed linker error (duplicate symbols for attitude conversion) Added comments to attitude elements conversions Fixed errors in quaternion order redefinition Minor changes
