+++ Tudat/Astrodynamics/BasicAstrodynamics/attitudeElementConversions.h
@@ -31,9 +31,6 @@
 #include <Eigen/Core>
 #include <Eigen/Geometry>
 
-#include "Tudat/Astrodynamics/BasicAstrodynamics/stateVectorIndices.h"
-#include "Tudat/Mathematics/BasicMathematics/mathematicalConstants.h"
-
 namespace tudat
 {
 
@@ -41,107 +38,40 @@ namespace orbital_element_conversions
 {
 
 //! Convert quaternions to modified Rodrigues parameters.
-Eigen::Vector4d convertQuaternionsToModifiedRodriguesParameterElements( const Eigen::Vector4d& quaternionElements )
-{
-    // Declare eventual output vector
-    Eigen::Vector4d convertedModifiedRodriguesParameterElements;
-
-    // Extract quaternions
-    double etaQuaternionParameter = quaternionElements( etaQuaternionIndex );
-
-    // Convert quaternions to modified Rodrigues parameters (or SMPR)
-    bool shadowFlag = etaQuaternionParameter < 0;
-    double conversionSign = shadowFlag ? - 1.0 : 1.0; // conversion is slightly different for SMRP and MRP
-    convertedModifiedRodriguesParameterElements.segment( sigma1ModifiedRodriguesParametersIndex, 3 ) = conversionSign *
-            quaternionElements.segment( epsilon1QuaternionIndex, 3 ) / ( 1 + conversionSign * etaQuaternionParameter );
-    convertedModifiedRodriguesParameterElements( shadowFlagModifiedRodriguesParametersIndex ) = shadowFlag ? 1.0 : 0.0;
-
-    // Give output
-    return convertedModifiedRodriguesParameterElements;
-}
+/*!
+ * Convert quaternions to modified Rodrigues parameters. The conversion is slightly different for modified Rodrigues parameters (MRP)
+ * than for shadow modified Rodrigues parameters (SMRP). This difference is embodied by conversionSign.
+ * \param quaternionElements Vector of quaternion elements.
+ * \return convertedModifiedRodriguesParameterElements Vector of (shadow) modified Rodrigues parameter elements.
+ */
+Eigen::Vector4d convertQuaternionsToModifiedRodriguesParameterElements( const Eigen::Vector4d& quaternionElements );
 
 //! Convert modified Rodrigues parameters to quaternions.
-Eigen::Vector4d convertModifiedRodriguesParametersToQuaternionElements( const Eigen::Vector4d& modifiedRodriguesParameterElements )
-{
-    // Declare eventual output vector
-    Eigen::Vector4d convertedQuaternionElements;
-
-    // Extract modified Rodrigues parameters
-    Eigen::Vector3d modifiedRodriguesParametersVector =
-            modifiedRodriguesParameterElements.segment( sigma1ModifiedRodriguesParametersIndex, 3 );
-
-    // Precompute often used variables
-    double modifiedRodriguesParametersMagnitudeSquared = modifiedRodriguesParametersVector.squaredNorm( );
-
-    // Convert modified Rodrigues parameters to quaternions
-    double conversionSign = ( int( modifiedRodriguesParameterElements( shadowFlagModifiedRodriguesParametersIndex ) ) == 1 ) ?
-                - 1.0 : 1.0; // converion is slightly different for SMRP and MRP
-    convertedQuaternionElements( etaQuaternionIndex ) = conversionSign * ( 1.0 - modifiedRodriguesParametersMagnitudeSquared ) /
-            ( 1.0 + modifiedRodriguesParametersMagnitudeSquared );
-    convertedQuaternionElements.segment( epsilon1QuaternionIndex, 3 ) = conversionSign *
-            2.0 / ( 1.0 + modifiedRodriguesParametersMagnitudeSquared ) *
-            modifiedRodriguesParametersVector;
-
-    // Give output
-    return convertedQuaternionElements;
-}
+/*!
+ * Convert modified Rodrigues parameters to quaternions. The conversion is slightly different for modified Rodrigues parameters (MRP)
+ * than for shadow modified Rodrigues parameters (SMRP). This difference is embodied by conversionSign.
+ * \param modifiedRodriguesParameterElements Vector of (shadow) modified Rodrigues parameters elements.
+ * \return convertedQuaternionElements Vector of quaternion elements.
+ */
+Eigen::Vector4d convertModifiedRodriguesParametersToQuaternionElements( const Eigen::Vector4d& modifiedRodriguesParameterElements );
 
 //! Convert quaternions to exponential map.
-Eigen::Vector3d convertQuaternionsToExponentialMapElements( const Eigen::Vector4d& quaternionElements )
-{
-    using mathematical_constants::PI;
-
-    // Define the tolerance of a singularity
-    const double singularityTolerance = 20.0 * std::numeric_limits< double >::epsilon( );
-
-    // Declare eventual output vector
-    Eigen::Vector3d convertedExponentialMapElements = quaternionElements.segment( epsilon1QuaternionIndex, 3 );
-
-    // Convert quaternions to exponential map (or SEM)
-    double exponentialMapMagnitude = 2.0 * std::acos( quaternionElements( etaQuaternionIndex ) );
-    if ( std::fabs( exponentialMapMagnitude ) < singularityTolerance )
-    {
-        convertedExponentialMapElements *= 48.0 / ( 24.0 - exponentialMapMagnitude * exponentialMapMagnitude );
-    }
-    else
-    {
-        convertedExponentialMapElements *= ( std::fabs( exponentialMapMagnitude ) > PI ) ?
-                    - ( 2.0 * PI - exponentialMapMagnitude ) / std::sin( 0.5 * exponentialMapMagnitude ) : // shadow exponential map
-                    exponentialMapMagnitude / std::sin( 0.5 * exponentialMapMagnitude ); // exponential map
-    }
-
-    // Give output
-    return convertedExponentialMapElements;
-}
+/*!
+ * Convert quaternions to exponential map. The conversion is the same for both exponential map (EM) and shadow
+ * exponential map (SEM).
+ * \param quaternionElements Vector of quaternion elements.
+ * \return convertedExponentialMapElements Vector of (shadow) exponential map elements.
+ */
+Eigen::Vector3d convertQuaternionsToExponentialMapElements( const Eigen::Vector4d& quaternionElements );
 
 //! Convert exponential map to quaternions.
-Eigen::Vector4d convertExponentialMapToQuaternionElements( const Eigen::Vector3d& exponentialMapElements )
-{
-    // Define the tolerance of a singularity
-    const double singularityTolerance = 20.0 * std::numeric_limits< double >::epsilon( );
-
-    // Declare eventual output vector
-    Eigen::Vector4d convertedQuaternionElements;
-
-    // Extract exponential map
-    double exponentialMapMagnitude = exponentialMapElements.norm( );
-
-    // Convert exponential map to quaternions
-    convertedQuaternionElements( etaQuaternionIndex ) = std::cos( 0.5 * exponentialMapMagnitude );
-    if ( std::fabs( exponentialMapMagnitude ) < singularityTolerance )
-    {
-        convertedQuaternionElements.segment( epsilon1QuaternionIndex, 3 ) =
-                exponentialMapElements * ( 0.5 - exponentialMapMagnitude * exponentialMapMagnitude / 48.0 );
-    }
-    else
-    {
-        convertedQuaternionElements.segment( epsilon1QuaternionIndex, 3 ) =
-                exponentialMapElements / exponentialMapMagnitude * std::sin( 0.5 * exponentialMapMagnitude );
-    }
-
-    // Give output
-    return convertedQuaternionElements;
-}
+/*!
+ * Convert exponential map to quaternions. The conversion is the same for both exponential map (EM) and shadow
+ * exponential map (SEM).
+ * \param exponentialMapElements Vector of (shadow) exponential map elements.
+ * \return convertedQuaternionElements Vector of quaternion elements.
+ */
+Eigen::Vector4d convertExponentialMapToQuaternionElements( const Eigen::Vector3d& exponentialMapElements );
 
 } // namespace orbital_element_conversions
