+++ Tudat/Astrodynamics/BasicAstrodynamics/unifiedStateModelModifiedRodriguesParameterElementConversions.h
@@ -13,8 +13,8 @@
  *
  */
 
-#ifndef TUDAT_UNIFIED_STATE_MODEL_MODIFIED_RODRIGUES_PARAMETERS_ELEMENT_CONVERSIONS_H
-#define TUDAT_UNIFIED_STATE_MODEL_MODIFIED_RODRIGUES_PARAMETERS_ELEMENT_CONVERSIONS_H
+#ifndef TUDAT_UNIFIED_STATE_MODEL_MODIFIED_RODRIGUES_PARAMETER_ELEMENT_CONVERSIONS_H
+#define TUDAT_UNIFIED_STATE_MODEL_MODIFIED_RODRIGUES_PARAMETER_ELEMENT_CONVERSIONS_H
 
 #include "Tudat/Basics/basicTypedefs.h"
 
@@ -129,4 +129,4 @@ Eigen::Matrix< double, 6, 1 > convertUnifiedStateModelModifiedRodriguesParameter
 
 } // close tudat
 
-#endif // TUDAT_UNIFIED_STATE_MODEL_MODIFIED_RODRIGUES_PARAMETERS_ELEMENT_CONVERSIONS_H
+#endif // TUDAT_UNIFIED_STATE_MODEL_MODIFIED_RODRIGUES_PARAMETER_ELEMENT_CONVERSIONS_H
