+++ Tudat/Astrodynamics/BasicAstrodynamics/unifiedStateModelQuaternionElementConversions.h
@@ -13,8 +13,8 @@
  *
  */
 
-#ifndef TUDAT_UNIFIED_STATE_MODEL_QUATERNIONS_ELEMENT_CONVERSIONS_H
-#define TUDAT_UNIFIED_STATE_MODEL_QUATERNIONS_ELEMENT_CONVERSIONS_H
+#ifndef TUDAT_UNIFIED_STATE_MODEL_QUATERNION_ELEMENT_CONVERSIONS_H
+#define TUDAT_UNIFIED_STATE_MODEL_QUATERNION_ELEMENT_CONVERSIONS_H
 
 #include "Tudat/Basics/basicTypedefs.h"
 
@@ -130,4 +130,4 @@ Eigen::Matrix< double, 6, 1 > convertUnifiedStateModelQuaternionsToCartesianElem
 
 } // close tudat
 
-#endif // TUDAT_UNIFIED_STATE_MODEL_QUATERNIONS_ELEMENT_CONVERSIONS_H
+#endif // TUDAT_UNIFIED_STATE_MODEL_QUATERNION_ELEMENT_CONVERSIONS_H
